+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -104,6 +104,11 @@ New check aliases
 Changes in existing checks
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- Improved :doc:`bugprone-use-after-move
+  <clang-tidy/checks/bugprone/use-after-move>` check to handle smart pointers
+  like any other objects allowing to detect more cases, previous behavior can
+  be restored by setting `IgnoreNonDerefSmartPtrs` option to `true`. 
+
 Removed checks
 ^^^^^^^^^^^^^^